+++ src/azure/resource-manager/authenticate.test.ts
@@ -1,7 +1,7 @@
 import { Recording } from '@jupiterone/integration-sdk-testing';
 
 import config from '../../../test/integrationInstanceConfig';
-import authenticate from './authenticate';
+import { validateResourceManagerInvocation } from './authenticate';
 import { setupAzureRecording } from '../../../test/helpers/recording';
 
 let recording: Recording;
@@ -10,34 +10,40 @@ afterEach(async () => {
   await recording.stop();
 });
 
-test('authenticate with subscription matching', async () => {
-  recording = setupAzureRecording({
-    directory: __dirname,
-    name: 'authenticate',
+describe('validateResourceManagerInvocation', () => {
+  test('authenticate with subscription matching', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'authenticate',
+    });
+    const credentials = await validateResourceManagerInvocation(config);
+    expect(credentials).toBeUndefined();
   });
-  const credentials = await authenticate(config);
-  expect(credentials).toBeDefined();
-});
 
-test('authenticate with subscription not matching', async () => {
-  recording = setupAzureRecording({
-    directory: __dirname,
-    name: 'authenticate bad subscription',
+  test('authenticate with subscription not matching', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'authenticate bad subscription',
+    });
+    await expect(
+      validateResourceManagerInvocation({ ...config, subscriptionId: 'junk' }),
+    ).rejects.toThrow(
+      'subscriptionId not found in tenant specified by directoryId',
+    );
   });
-  await expect(
-    authenticate({ ...config, subscriptionId: 'junk' }),
-  ).rejects.toThrow(/not found in tenant/);
-});
 
-test('authenticate with no subscription', async () => {
-  recording = setupAzureRecording({
-    directory: __dirname,
-    name: 'authenticate no subscription',
+  test('authenticate with no subscription', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'authenticate no subscription',
+    });
+    await expect(
+      validateResourceManagerInvocation({
+        ...config,
+        subscriptionId: undefined,
+      }),
+    ).rejects.toThrow(
+      'Cannot use Azure Resource Manager APIs without subscriptionId',
+    );
   });
-  await expect(
-    authenticate({
-      ...config,
-      subscriptionId: undefined,
-    }),
-  ).rejects.toThrow(/without/);
 });